+++ doc/changelog.rst
@@ -3,6 +3,7 @@
 Release Notes
 =============
 
+- :feature:`656` You can now choose if you want to compare the median of review scores or the average/mean.
 - :feature:`313` Organisers can now create custom proposal and speaker exports (as either CSV or JSON), including any data they need.
 - :support:`-` If you use PostgreSQL, please make sure to be on at least PostgreSQL 10.
 - :feature:`-` pretalx will no longer send emails to localhost addresses, as those are used for deleted users. Please get in touch if this is a problem for you.